major restructuring; code clean up; refactoring
