+++ src/main/resources/application.conf
@@ -6,6 +6,8 @@
 akka {
   loglevel = INFO
   heartbeat-interval = 5s
+  log-dead-letters = 10
+  log-dead-letters-during-shutdown = on
 }
 
 spray.can.server {
