+++ src/main/scala/edu/ucsc/vesper/http/api/AsyncSupport.scala
@@ -1,9 +1,11 @@
-package edu.ucsc.vesper.http.core
+package edu.ucsc.vesper.http.api
+
+import java.util.concurrent.TimeUnit._
 
 import akka.pattern.AskSupport
 import akka.util.Timeout
+
 import scala.concurrent.duration.Duration
-import java.util.concurrent.TimeUnit._
 
 /**
  * @author hsanchez@cs.ucsc.edu (Huascar A. Sanchez)
