+++ src/main/scala/edu/ucsc/vesper/http/api/KiwiActor.scala
@@ -5,7 +5,7 @@ import akka.actor.Actor
 /**
  * @author hsanchez@cs.ucsc.edu (Huascar A. Sanchez)
  */
-class VesperinApi extends Actor with Vesperin {
+class KiwiActor extends Actor with Kiwi {
   // the HttpService trait defines only one abstract member, which
   // connects the services environment to the enclosing actor or test
   override def actorRefFactory = context
@@ -13,5 +13,5 @@ class VesperinApi extends Actor with Vesperin {
   // this actor only runs our route, but you could add
   // other things here, like request stream processing
   // or timeout handling
-  override def receive = runRoute(vesperRoutes)
+  override def receive = runRoute(routes)
 }
